# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nuclom API",
    "version": "1.0.0",
    "description": "Nuclom is a video collaboration platform that helps teams organize, share, and collaborate on video content. This API provides programmatic access to videos, organizations, series, comments, and more.",
    "contact": {
      "name": "Nuclom Support",
      "url": "https://nuclom.com/support"
    },
    "license": {
      "name": "Elastic License 2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    }
  },
  "servers": [
    {
      "url": "https://nuclom.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List organizations",
        "operationId": "listOrganization",
        "description": "Get organizations the current user is a member of",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of organizations",
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "UUID": {
                      "type": "string",
                      "description": "a Universally Unique Identifier",
                      "format": "uuid",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                    }
                  },
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "slug",
                      "description",
                      "logo",
                      "createdAt",
                      "updatedAt"
                    ],
                    "properties": {
                      "id": {
                        "$ref": "#/$defs/UUID"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "description": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "logo": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    },
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create organization",
        "operationId": "createOrganization",
        "description": "Create a new organization",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Create organization data",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "a string that will be trimmed"
                  },
                  "slug": {
                    "type": "string",
                    "description": "a string matching the pattern ^[a-z0-9-]+$",
                    "title": "maxLength(50)",
                    "minLength": 1,
                    "maxLength": 50,
                    "pattern": "^[a-z0-9-]+$"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string",
                            "description": "a string that will be trimmed"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "logo": {
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "UUID": {
                      "type": "string",
                      "description": "a Universally Unique Identifier",
                      "format": "uuid",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                    }
                  },
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "logo",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "$ref": "#/$defs/UUID"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "logo": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/organizations/{id}/roles": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get rol",
        "operationId": "getOrganizationRol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create rol",
        "operationId": "createOrganizationRol",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/organizations/{id}/settings": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get setting",
        "operationId": "getOrganizationSetting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Patch setting",
        "operationId": "patchOrganizationSetting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication token from login or API key"
      }
    }
  }
}